refactored oauth endpoints and adds rate limiting on the client endpoint
